working on sequencer
